+++ NEWS.md
@@ -157,3 +157,4 @@
 * Issue a warning if duplicate antigen or serum names are found when creating a map
 
 # Racmacs 1.1.34
+* Add support for getting and setting serum species with `srSpecies()`.